Updated PropertyInfo conventions to use PropertyMaps instead.
Now ForMember statements will be used to compare the source and destination if applied.  Default conventions still apply.
Examples:
-ForMember going to different property which name won't match will use that as an equivalent value
-ForMember using something other than just a property to match (sub properties, arithmetic, sub functions)

Issues with using it now (Don't have a scenario why someone might run into this but it's still possible)
Equality assumes both values are of the same type and simple types
-If complex type equality won't work because it will default to equals
-If properties are different and have another type map, will get an exception and mapping will default to can't convert
Way to fix would be to add recursion into the equivalency expression so that if source and destination don't have same type, return a new equivalency statement to be added to the parent comparison