Work around xunit issue which drops everything after dashes.
xunit has a problem with rendering exception messages that inclued lines
of dashes in the message. It silently drops everything after the dashes.
This is documented at https://github.com/xunit/xunit/issues/149

This commit adds a slightly more verbose wording to the message
and updates the dashes to a blank line.