feature: TypePair ignore EntityFramework proxy types when getting type from source instance
When specifying an instance to be mapped to the IMapper.Map the source.GetType() takes precedence over the specified sourceType. If source is an instance of an EntityFramework proxy object and there is a defined map for the POCO the proxy is sub classing, the custom map is not used as the sourceType is the EF proxy not the POCO type.